-Fix all warnings (except for quaternion)
-Fix compiler error in Release

+++ Outerrain/Include/cameraOrbiter.h
@@ -23,17 +23,17 @@ public:
 	void Frame(const float width, const float height, const float z, const float fov, Vector3& dO, Vector3& dx, Vector3& dy) const;
 	 
 	Vector3 Position() const;
-	void SetFrameWidth(float w);
-	void SetFrameHeight(float h);
-	float FrameWidth() const;
-	float FrameHeight() const;
+	void SetFrameWidth(int w);
+	void SetFrameHeight(int h);
+	int FrameWidth() const;
+	int FrameHeight() const;
 
 protected:
 	Vector3 center;
 	Vector2 position;
 	Vector2 rotation;
 	float size;
-	float frameWidth, frameHeight;
+	int frameWidth, frameHeight;
 	float zNear, zFar;
 };
